Dynamic routing and Records
The record view now populates based on the record id passed into the
url and dynamically picks the record to be show. Currently it is
pulling off a static set of records (well a set of one) but largely it
works. Also added the bare component of the add record tab.